backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "gunicorn>=23.0.0",
    "psycopg2>=2.9.10",
    "whitenoise>=6.9.0",
]
